db연결 테스트
