+++ app/models/member.py
@@ -1,8 +1,9 @@
 from sqlalchemy import Column, Integer, String
-from app.test_db import Base
+from app.services.mysql_service import Base
 
 class Member(Base):
     __tablename__ = "member"
+    
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String, nullable=False)
-    email = Column(String, nullable=False)
+    email = Column(String, nullable=False, unique=True)
