+++ app/test_db.py
@@ -1,30 +1,8 @@
-import os
-from dotenv import load_dotenv
-from sqlalchemy import create_engine
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import sessionmaker
+from app.services.mysql_service import Base, engine
+from app.models.member import Member  # Member 모델 임포트
 
-# .env 파일 로드
-load_dotenv()
+# 테이블 생성 (삭제 없이 생성만)
+print("새 테이블 생성 중...")
+Base.metadata.create_all(bind=engine)
 
-# 환경 변수에서 DATABASE_URL 가져오기
-DATABASE_URL = os.getenv("DATABASE_URL")
-
-if not DATABASE_URL:
-    raise ValueError("DATABASE_URL 환경 변수가 설정되지 않았습니다.")
-
-# 데이터베이스 연결 설정
-engine = create_engine(DATABASE_URL, connect_args={"check_same_thread": False})
-SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-Base = declarative_base()
-
-# 데이터베이스 세션 관리 함수
-def get_db():
-    """
-    요청마다 독립적인 데이터베이스 세션을 생성하여 관리합니다.
-    """
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
+print("MySQL 데이터베이스 테이블 생성 완료")
